Catching exceptions
Hi
There seem to be an issue with the lib. Some users had it with the app we created, and i was able to reproduce it recently.
When the device has no internet connection, the request will fail. There is a request callback, and a response callback, both are async. The ResponseCallback exceptions are catched, but the RequestCallback exceptions are not. Even if you catch all exceptions around the stathat library calls, the fact that the callbacks are run in another thread makes the exception non catched and the app will crash.
Hope this helps
Jerome